<refactor/structure>: seperate method service in .controller package to .service package (#19)
